tea/requirements.txt
Django==3.2.5
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
Markdown==3.3.4
celery
